requirements.txt
gitpython>=3.1.30
matplotlib>=3.3
opencv-python>=4.1.1
Pillow>=7.1.2
psutil
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
thop>=0.1.1
torch>=2.0.1
torchvision>=0.15.2
tqdm>=4.64.0
ultralytics>=8.0.111
pandas>=1.1.4
seaborn>=0.11.0
onnx>=1.14.0
onnxruntime>=1.15.1
streamlit
streamlit_option_menu
folium
plotly
geopy